Implement jwt & passport

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { TaskRepository } from './tasks.repository';
 import { Task } from './tasks.entity';
 import { TaskStatus } from './tasks.enum';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -14,33 +15,33 @@ export class TasksService {
 
     }
 
-    async getTasks(searchTask: SearchTaskDto): Promise<Task[]> {
-        return await this.taskRepository.getTasks(searchTask);
+    async getTasks(searchTask: SearchTaskDto, user: User): Promise<Task[]> {
+        return await this.taskRepository.getTasks(searchTask, user);
     }
 
-    async getTaskById(id: string): Promise<Task> {
-        const task = await this.taskRepository.findOne(id);
+    async getTaskById(id: string, user?: User): Promise<Task> {
+        const task = await this.taskRepository.findOne({id, userId: user.id});
         if (!task) {
             throw new NotFoundException(`Task with ID ${id} not found`);
         }
         return task;
     }
 
-    async deleteTask(id: string): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(id: string, user: User): Promise<void> {
+        const result = await this.taskRepository.delete({ id, userId: user.id});
         if (result.affected === 0) {
             throw new NotFoundException(`Task with ID ${id} not found`);
         }
     }
 
-    async updateStatus(id: string, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateStatus(id: string, status: TaskStatus, user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(createTaskDto);
+    async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDto, user);
     }
 }
